Fixes Instagram and YouTube login flow
Addresses issues with the Instagram and YouTube login flow by adding a fallback mechanism to handle cases where the popup window is closed before the code is received. This ensures that the login loading state is correctly updated and prevents potential errors.

Also renames pages to be in index files.

Updates the Dockerfile to expose the correct port.